Migrate some old Java constructs
- Remove unboxing
- Replace lambdas to method references
- Remove some redundancies

+++ qosservice/impl/src/main/java/org/opendaylight/netvirt/qosservice/QosAlertManager.java
@@ -147,7 +147,7 @@ public final class QosAlertManager implements Runnable {
         alertEnabled = defaultConfig.isQosAlertEnabled();
         pollInterval = defaultConfig.getQosAlertPollInterval();
 
-        alertThresholdSupplier.set(defaultConfig.getQosDropPacketThreshold().shortValue());
+        alertThresholdSupplier.set(defaultConfig.getQosDropPacketThreshold());
     }
 
     public void setQosalertConfig(QosalertConfig config) {
